Relocate sstring class
