+++ c/frame/repr/text_column.h
@@ -24,7 +24,7 @@
 #include <iostream>
 #include <string>
 #include <vector>
-#include "frame/repr/sstring.h"
+#include "utils/terminal/sstring.h"
 #include "utils/terminal/terminal.h"
 #include "utils/terminal/terminal_stream.h"
 #include "column.h"
