+++ c/utils/terminal/sstring.h
